Add support for adding annotations to OCI indexes
OCI indexes support annotations and many tools/registries recommend and
use the standard OCI org annotations to improve metadata visibility of
images. With these changes manifest-tool will properly push annotations
either as command line options or included in the YAML (the example
YAML in the repo is updated to show this as well).

Since the official Docker v2 manifestList schema does not support
annotations at the manifestList level (only on the manifests themselves)
this tool will error out if you try and submit annotations but don't
use the "--type oci" flag on the command line.

Signed-off-by: Phil Estes <estesp@gmail.com>